pyproject.toml
[project]
name = "evetradeanalyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "aiohttp>=3.11.11",
    "django>=5.1.4",
    "django-background-tasks>=1.2.8",
    "pandas>=2.2.3",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

